Change gain setting type and update thread print info
Changed the `set_gain` function in `_camera_settings_func.py` to receive a float parameter instead of an integer for more precision on camera settings. In `tests_threading.py`, added current time in `print` function for more informative output regarding thread testing. Both changes enhance precision and give more detailed information.

+++ src/gui/_camera_settings_func.py
@@ -28,7 +28,7 @@ def set_gain(self, num):
         self.show_camera_error()
         return
     
-    self.cam[num].set_gain(int(self.gain[num].get()))
+    self.cam[num].set_gain(float(self.gain[num].get()))
     get_frame_rate_list(self, num)
